scripts/zotero/requirements.txt
pyzotero
python-dotenv
requests
